Removed speed from animation Playback.elapsed
I have migrated the places where we set speeds for an animation to call `updateSpeed` so there is no more real need always passing the animation speed to Playback.elapsed since we are just going to use whatever is set in the playback instance anyways.  The changes in PR entirely remove passing in `speed` and instead you will need to either call `setSpeed` in the cases where we want to set the speed of an animation before it starts or `updateSpeed` after the animation starts.