Converting `scenic/math` into a monorepo package
Im working on splitting up different independent components in scenic to be separate modules. To do this, Im converting scenic to be a monorepo.  The first thing in scenic Im converting to a module is math to help me start setting up the infrastructure to support monorepos with separate modules. All the different modules will be organized under the `packages` directory as an npm workspace.

There are several unanwered questions that I need to work thru:

1. The build for each module in packages. This includes correctly building bundles excluding inter dependencies on other scenic modules.  For example, we dont want the math module to appear in each scenic module that imports it. This will likely require updates to bit-bundler.
2. Figure the deploy process for each module.

The next module I will be migrating is scene.