Configured eslint and babelrc with shared configurations
As we transition to monorepo with individual package modules in packages, we needed to setup a way to share configurations for babel and eslint.  The tool for the job is symlinks so that we can update one central file and it be reflected in each module.

Also added a README.md with details about the packages directory