Transitioned animation to @scenic/animation
We have been working on transitioning all the different part of scenic to a monorepo with separate packages under the `@scenic` npm org.  Previously, I migrated math, renderer, and others to `@scenic/math`, `@scenic/renderer` and so on.  The changes in this commit transition animation to `@scenic/animation`.

The next changes will be to convert `format`.