Added noop test in package.json for @scenic/dom
I had configured the test script for the @scenic/dom package to call `jest`, but there are no tests to run, so that was call was failing everytime we had to run all the tests in all the other packages.  So I changed the `test` script to `echo no test`.