Moved fbx and obj file format parser to @scenic/fbx and @scenic/obj
We have been working on switching all the different parts of scenic to be a monorepo with separate npm packages for all the different bits. I have migrated @scenic/math, @scenic/renderer, @scenic/dom, and so on.  The changes in this commit migrate the fbx and obj file format parsers into their own separate packages.

For organizational reasons I have created a single `encoding` directory with an `fbx` and `obj` package in it; two separate packages.  Other file formats that we add support for will just be added in the `encoding` directory.