Moved scene code to @scenic/scene package
We have been working on transtioning scenic code to separate libraries under the @scenic npm org. Previously I transitioned @scenic/math, @scenic/dom, and others. The changes in this commit transition @scenic/scene, which is the largest and last of the large switches that need to be made.  To make this smooth I moved the node helper function such as findParentByType and findChildrenByType to be static functions in the Node class. Other than that, things are pretty much left untouched.

Next is figure out what to do with the `utils` directory!