Converted packages/utils and the editor to proper workspace packages
This is a huge change.  At least it looks like it. Initially this was supposed to only be converting packages/utils to a proper workspace package.  But I accidentally also added all the changes to also convert the editor into a package, which is moving a lot of files around.  Sigh.  Anyways - now we have a proper apps workspace as well.